requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coverage==5.3
coveralls==2.1.2
cryptography==2.9.2
Django==2.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
flake8==3.8.4
idna==2.10
importlib-metadata==2.0.0
libcst==0.3.13
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
zipp==3.3.1